backend/requirements.txt
fastapi
uvicorn
playwright
openai
python-dotenv
pandas
beautifulsoup4
httpx